yolo-service/requirements_yolo.txt
fastapi
uvicorn
ultralytics
pillow
motor
pyyaml
httpx
mlflow
python-multipart